Linux/MQTT/publisher/requirements.txt
paho-mqtt==1.6.1
RPi.GPIO==0.7.1a4
Adafruit-DHT==1.4.0
